+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 
 - A new module `frequenz.client.dispatch.test` has been added, providing a fake Service and Client as well as a `DispatchGenerator` to generate `Dispatch` instances filled with random data.
 - The `DispatchGenerator.generate_dispatch` method now accepts a `microgrid_id` parameter to generate `Dispatch` instances with a specific microgrid ID.
+- The `Client.create()` method now returns the newly created `Dispatch` object.
 
 ## Bug Fixes